+++ var/starter_kits/italiagov/webpack.paths.js
@@ -1,0 +1,11 @@
+const path = require('path');
+
+module.exports = {
+  // Source files
+  src: path.resolve(__dirname, 'src'),
+
+  // Destination build files
+  build: path.resolve(__dirname, 'assets'),
+
+  modules: path.resolve(__dirname, 'node_modules'),
+}